fix: move utils.dart to the utils folder
